+++ firmware/lib/drivers/wpump_converter.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
 
-uint16_t wpump_converter_convert_mL_to_ms(uint16_t number_in_mL);
+uint32_t wpump_converter_convert_mL_to_ms(uint32_t number_in_mL);
